requirements.txt
Flask==2.0.1
Flask-Uploads
pytesseract==0.3.10
Pillow
